+++ pd-feature-tests/src/test/java/org/uh/attx/platform/test/TestRunner.java
@@ -3,18 +3,17 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package org.uh.attx.platform.testing;
+package org.uh.attx.platform.test;
 
 import cucumber.api.CucumberOptions;
 import cucumber.api.junit.Cucumber;
 import org.junit.runner.RunWith;
 
 /**
- *
  * @author jkesanie
  */
 @RunWith(Cucumber.class)
-@CucumberOptions(plugin = {"pretty"}, features="classpath:features", tags = {"~@ignore"}, glue={"classpath:org.uh.attx.platform.testing", "src/test/resources/python/steps"})
+@CucumberOptions(plugin = {"pretty"}, features = "classpath:features", tags = {"~@ignore"}, glue = {"classpath:org.uh.attx.platform.test"})
 public class TestRunner {
-    
+
 }
