+++ pd-feature-tests/src/test/java/org/uh/attx/platform/test/UVPluginsSteps.java
@@ -3,24 +3,23 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package org.uh.attx.platform.testing;
+package org.uh.attx.platform.test;
 
 import com.mashape.unirest.http.HttpResponse;
 import com.mashape.unirest.http.JsonNode;
 import com.mashape.unirest.http.Unirest;
-import com.mashape.unirest.request.GetRequest;
 import cucumber.api.java8.En;
+
 import java.io.File;
-import junit.framework.Assert;
+
 import junit.framework.TestCase;
 import org.skyscreamer.jsonassert.JSONAssert;
 import org.skyscreamer.jsonassert.JSONCompareMode;
 
 /**
- *
  * @author jkesanie
  */
-public class UVPluginsSteps  implements En {
+public class UVPluginsSteps implements En {
     private PlatformServices s = new PlatformServices();
 
     private final String API_USERNAME = "master";
@@ -31,12 +30,10 @@ public class UVPluginsSteps  implements En {
         Given("^that platform is up and running$", () -> {
             // Write code here that turns the phrase above into concrete actions
             TestCase.assertTrue(true);
-
         });
 
         Then("^UV should contain uv-metadata plugin$", () -> {
             try {
-
                 // using the UV rest API and trying to add the plugin again. Should response with an error.
                 HttpResponse<JsonNode> response = Unirest.post(s.getUV() + "/master/api/1/import/dpu/jar")
                         .header("accept", "application/json")
@@ -46,14 +43,12 @@ public class UVPluginsSteps  implements En {
                         .basicAuth(API_USERNAME, API_PASSWORD)
                         .asJson();
 
-                System.out.println("**"  + response.getBody().toString());
+                System.out.println("**" + response.getBody().toString());
                 JSONAssert.assertEquals("{\"errorMessage\": \"DPU with this name already exists.\"}", response.getBody().toString(), JSONCompareMode.LENIENT);
 
                 TestCase.assertEquals(400, response.getStatus());
 
-
-
-            } catch(Exception ex ) {
+            } catch (Exception ex) {
                 TestCase.fail(ex.getMessage());
             }
         });
